Issue 130: added options to runNode like openPorts and runScript; added ability to specify alternate endpoint in *ContextBuilderFactory; updated ant tasks to use new node features; Issue 112: normalized vcloud clients
git-svn-id: http://jclouds.googlecode.com/svn/trunk@2712 3d8758e0-26b5-11de-8745-db77d3ebf521

+++ vcloud/terremark/src/main/java/org/jclouds/vcloud/terremark/domain/ComputeOptions.java
@@ -29,10 +29,10 @@ package org.jclouds.vcloud.terremark.domain;
  */
 public class ComputeOptions implements Comparable<ComputeOptions> {
    private final int processorCount;
-   private final long memory;
+   private final int memory;
    private final float costPerHour;
 
-   public ComputeOptions(int processorCount, long memory, float costPerHour) {
+   public ComputeOptions(int processorCount, int memory, float costPerHour) {
       this.processorCount = processorCount;
       this.memory = memory;
       this.costPerHour = costPerHour;
@@ -42,7 +42,7 @@ public class ComputeOptions implements Comparable<ComputeOptions> {
       return processorCount;
    }
 
-   public long getMemory() {
+   public int getMemory() {
       return memory;
    }
